refactor(Test): 패키지를 프로덕션과 같도록 수정
